Eina_Debug: reconnect session if disconnected
if session disconnected we will init the opcodes array ( opcodes number
maybe will be different next time we will connect ), tell the
module that send the opcodes that they are not ready/ not set ( using the status cb).
If we will reconnect we will re register all opcodes and tell the
modules the their opcodes are ready.