Improve remote shell connection
Because some shells echoes the sent requests, we need a way to distinct
between a request sent to the shell bridge and the response.
A magic number is now expected in the packets sent by the shell bridge.
The receive function will drop bytes until this magic number is not
found.

In addition, WAIT token helps synchronization with the scripts commands.
As some commands sends back data (ssh shell) and some don't, we need a
way to indicate the script handler if a new command has to be sent.

Only no more command has to be executed, the greet packet is sent.