Ethumb: make the emotion plugin correctly use the edje frame if requested.
The emotion plugin was lacking the ability to use the edje frame as expected,
this patch add this ability, so now video thumbnail works like the normal ones.

This commit is a bit of a new feature, but I consider this a @fix as now
the ethumb_frame_set() API works as expected, also the implementation
is so simple that I think it can go in without any problems.