manager: introduce borders to the nodes
a border is just a list of relations to a other node. So each direction
of a node can have multiple relations to other nodes. To still keep the
way of "reverting" a previous operation, the manager looks for the most
recent focused element in of all the border relations. So a user can
move right left right left and end up and the same node than before.