ecore_con: don't use EAI_SYSTEM when not defined
Windows has no support for EAI_SYSTEM. It seems not to have any
equivalent. Fortunately, in our context, EAI_SYSTEM was used
systematically in the same error checking routine. So it has been
factorized to be written once. If we don't have EAI_SYSTEM, the
checking routine will try to do without.